requirements.txt
anyio==3.6.1
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dnspython==2.2.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pymongo==4.2.0
python-telegram-bot==20.0a2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.2
tzdata==2022.2
tzlocal==4.2
